DOC: Documented all the components
