+++ components/Button/ExpandCollapseButton.tsx
@@ -7,6 +7,15 @@ interface ExpandCollapseButtonProps {
   className?: string;
 }
 
+/**
+ * Button that allows the user to expand or collapse content.
+ * This is used to show more or less information.
+ *
+ * @param isExpanded If the content is expanded or not
+ * @param onToggle Function to toggle the content
+ * @param className Additional classes for the button
+ * @returns Button to expand or collapse content
+ */
 const ExpandCollapseButton: React.FC<ExpandCollapseButtonProps> = ({
   isExpanded,
   onToggle,
