+++ components/Inputs/RadioButton.tsx
@@ -17,14 +17,15 @@ interface RadioButtonProps {
 /**
  * Radio Button component which allows the user to select one option.
  * The radio button is checked when the user clicks on it.
- * @param id (string): The id of the radio button.
- * @param name (string): The name of the radio button.
- * @param value (string): The value of the radio button.
- * @param checked (boolean): Whether the radio button is checked or not.
- * @param onChange (function): A function that handles the onChange event.
- * @param label (string): The label of the radio button.
- * @param className (string): The custom classes to be applied to the radio button.
- * @returns (JSX.Element): A radio button component.
+ *
+ * @param id (The ID of the radio button.
+ * @param name The name of the radio button.
+ * @param value The value of the radio button.
+ * @param checked Whether the radio button is checked or not.
+ * @param onChange A function that handles the onChange event.
+ * @param label The label of the radio button.
+ * @param className The custom classes to be applied to the radio button.
+ * @returns A radio button component.
  */
 const RadioButton: React.FC<RadioButtonProps> = ({
   id,
@@ -36,19 +37,19 @@ const RadioButton: React.FC<RadioButtonProps> = ({
   onChange,
 }) => {
   const [hovered, setHovered] = useState(false);
-  const isMounted = useIsMounted();
+  const isMounted: boolean = useIsMounted();
 
   if (!isMounted) {
     return null;
   }
 
-  const borderColor = checked
+  const borderColor: string = checked
     ? "border-red-500 dark:border-red-900"
     : hovered
-      ? "border-neutral-300 dark:border-red-950"
-      : "border-neutral-200 dark:border-neutral-800";
+    ? "border-neutral-300 dark:border-red-950"
+    : "border-neutral-200 dark:border-neutral-800";
 
-  const combinedClassName = twMerge(
+  const combinedClassName: string = twMerge(
     `
       border-2 ${borderColor}
       bg-neutral-200 dark:bg-neutral-800
@@ -59,11 +60,11 @@ const RadioButton: React.FC<RadioButtonProps> = ({
       my-2 
       flex items-center cursor-pointer
     `,
-    className,
+    className
   );
 
   // Add the font weight conditionally based on the checked state
-  const labelClassName = checked ? "font-bold" : "";
+  const labelClassName: string = checked ? "font-bold" : "";
 
   return (
     <label
