+++ components/MaterialLists/CertificatesList.tsx
@@ -5,6 +5,14 @@ import Grid from "@/components/UI/Grid";
 import MaterialListProps from "@/interfaces/props/MaterialListProps";
 import React from "react";
 
+/**
+ * List of certificates grouped by category to be displayed section by section.
+ * Each section contains a title and a list of certificates.
+ * If there are no certificates to display, a message is shown.
+ *
+ * @param groupedCertificates List of certificates grouped by category to be displayed section by section
+ * @returns A list of certificates grouped by category
+ */
 const CertificatesList: React.FC<MaterialListProps> = ({
   groupedMaterial: groupedCertificates,
 }) => {
