+++ components/Navbar/NavbarOverlay.tsx
@@ -15,10 +15,11 @@ interface OverlayProps {
  * This displays the links to other pages when the hamburger menu is clicked.
  * It allows the user to navigate to other pages.
  * It is necessary on mobile devices because the navbar is hidden on mobile devices.
- * @param isOpen (boolean) - Whether the overlay is open or not
- * @param toggle (function) - Function to toggle the overlay
- * @param items (NavItem[]) - Array of NavItem objects to display in the overlay
- * @returns (JSX.Element) - Overlay component
+ *
+ * @param isOpen Whether the overlay is open or not
+ * @param toggle Function to toggle the overlay
+ * @param items Array of NavItem objects to display in the overlay
+ * @returns Overlay component
  */
 const NavbarOverlay: React.FC<OverlayProps> = ({ isOpen, toggle, items }) => {
   const isMobile = useMediaQuery("(max-width: 768px)");
