+++ components/Tags/SkillTag.tsx
@@ -20,6 +20,15 @@ interface TagProps {
   hide?: boolean;
 }
 
+/**
+ * Tag component for displaying a skill.
+ * This component can be clicked to navigate to the skill's material if the skill has associated material.
+ * Associated material are projects, blogs, certifications, etc.
+ *
+ * @param skillKey The key of the skill to be displayed
+ * @param hide Whether to hide the tag or not
+ * @returns A tag with the name of the skill
+ */
 const SkillTag: React.FC<TagProps> = ({ skillKey, hide }) => {
   const currentPath: string = usePathname();
   const skill: SkillInterface = skillDatabase[skillKey];
