+++ components/Text/HeadingThree.tsx
@@ -7,8 +7,8 @@ interface HeadingThreeProps {
 /**
  * Heading 3 text component shown at the top of each section but under main headings.
  *
- * @param title (string): heading to be displayed for each section
- * @returns (JSX.Element): subheading component
+ * @param title Heading to be displayed for each section
+ * @returns Subheading component
  */
 const HeadingThree: React.FC<HeadingThreeProps> = ({ title }) => {
   return <h2 className="text-2xl font-bold mb-6 mt-6">{title}</h2>;
