+++ components/TextLoop/TypeWriterTextLoop.tsx
@@ -9,9 +9,10 @@ type TypeWriterTextLoopProps = {
 /**
  * Loops through an array of strings and displays them one by one.
  * Uses a type writer effect to loop through the text.
- * @param className (string) - Class name to apply to the span element
- * @param loopItems (string[]) - Array of strings to loop through
- * @returns (JSX.Element): A span element that loops through the given array of strings
+ *
+ * @param className Class name to apply to the span element
+ * @param loopItems Array of strings to loop through
+ * @returns A span element that loops through the given array of strings
  */
 const TypeWriterTextLoop: React.FC<TypeWriterTextLoopProps> = ({
   className,
